+++ src/search_doc.h
@@ -15,9 +15,9 @@
 #include <string>
 #include <vector>
 
-#include "src/document.h"
 #include "src/file_root/file_root.h"
 #include "src/repository.h"
+#include "src/value_object/document.h"
 namespace search_doc::api {
 
 class SearchDoc {
